refactor(web): split API transport tests
